backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "fastembed>=0.7.0",
    "keras>=3.10.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "tensorflow>=2.19.0",
    "openai>=1.82.0",
    "qdrant-client>=1.14.2",
]
